+++ src/MooVC.Tests/Hosting/ThreadSafeHostedServiceTests/WhenStartAsyncIsCalled.cs
@@ -5,58 +5,58 @@ using Microsoft.Extensions.Logging;
 
 public sealed class WhenStartAsyncIsCalled
 {
-    private readonly ThreadSafeHostedService host;
-    private readonly IHostedService service;
+    private readonly ThreadSafeHostedService _host;
+    private readonly IHostedService _service;
 
     public WhenStartAsyncIsCalled()
     {
-        service = Substitute.For<IHostedService>();
-        host = new ThreadSafeHostedService(Substitute.For<ILogger<ThreadSafeHostedService>>(), [service]);
+        _service = Substitute.For<IHostedService>();
+        _host = new ThreadSafeHostedService(Substitute.For<ILogger<ThreadSafeHostedService>>(), [_service]);
     }
 
     [Fact]
     public async Task GivenAStoppedHostThenTheServiceStarts()
     {
         // Act
-        await host.StartAsync(CancellationToken.None);
+        await _host.StartAsync(CancellationToken.None);
 
         // Assert
-        await service.Received(1).StartAsync(Arg.Any<CancellationToken>());
+        await _service.Received(1).StartAsync(Arg.Any<CancellationToken>());
     }
 
     [Fact]
     public async Task GivenAStartedHostThenTheServiceIsNotStartedASecondTime()
     {
         // Arrange
-        await host.StartAsync(CancellationToken.None);
+        await _host.StartAsync(CancellationToken.None);
 
         // Act
-        await host.StartAsync(CancellationToken.None);
+        await _host.StartAsync(CancellationToken.None);
 
         // Assert
-        await service.Received(1).StartAsync(Arg.Any<CancellationToken>());
+        await _service.Received(1).StartAsync(Arg.Any<CancellationToken>());
     }
 
     [Fact]
     public async Task GivenARestartThenTheServiceIsStartedTheSecondTime()
     {
         // Arrange
-        await host.StartAsync(CancellationToken.None);
-        await host.StopAsync(CancellationToken.None);
+        await _host.StartAsync(CancellationToken.None);
+        await _host.StopAsync(CancellationToken.None);
 
         // Act
-        await host.StartAsync(CancellationToken.None);
+        await _host.StartAsync(CancellationToken.None);
 
         // Assert
-        await service.Received(2).StartAsync(Arg.Any<CancellationToken>());
-        await service.Received(1).StopAsync(Arg.Any<CancellationToken>());
+        await _service.Received(2).StartAsync(Arg.Any<CancellationToken>());
+        await _service.Received(1).StopAsync(Arg.Any<CancellationToken>());
     }
 
     [Fact]
     public void GivenAStoppedHostStartAsyncIsNotCalled()
     {
         // Assert
-        _ = service.DidNotReceive().StartAsync(Arg.Any<CancellationToken>());
+        _ = _service.DidNotReceive().StartAsync(Arg.Any<CancellationToken>());
     }
 
     [Fact]
@@ -64,14 +64,14 @@ public sealed class WhenStartAsyncIsCalled
     {
         // Arrange
         var expected = new InvalidOperationException("Service failed to start.");
-        service.When(service => service.StartAsync(Arg.Any<CancellationToken>())).Do(_ => throw expected);
+        _service.When(service => service.StartAsync(Arg.Any<CancellationToken>())).Do(_ => throw expected);
 
         // Act & Assert
-        AggregateException actual = await Assert.ThrowsAsync<AggregateException>(() => host.StartAsync(CancellationToken.None));
+        AggregateException actual = await Assert.ThrowsAsync<AggregateException>(() => _host.StartAsync(CancellationToken.None));
 
         _ = actual.InnerException.Should().Be(expected);
 
-        await service.Received(1).StopAsync(Arg.Any<CancellationToken>());
+        await _service.Received(1).StopAsync(Arg.Any<CancellationToken>());
     }
 
     [Fact]
@@ -80,15 +80,15 @@ public sealed class WhenStartAsyncIsCalled
         // Arrange
         var start = new InvalidOperationException("Service failed to start.");
         var stop = new InvalidOperationException("Service failed to stop.");
-        service.When(service => service.StartAsync(Arg.Any<CancellationToken>())).Do(_ => throw start);
-        service.When(services => service.StopAsync(Arg.Any<CancellationToken>())).Do(_ => throw stop);
+        _service.When(service => service.StartAsync(Arg.Any<CancellationToken>())).Do(_ => throw start);
+        _service.When(services => _service.StopAsync(Arg.Any<CancellationToken>())).Do(_ => throw stop);
 
         // Act & Assert
-        AggregateException actual = await Assert.ThrowsAsync<AggregateException>(() => host.StartAsync(CancellationToken.None));
+        AggregateException actual = await Assert.ThrowsAsync<AggregateException>(() => _host.StartAsync(CancellationToken.None));
 
         _ = actual.InnerException.Should().Be(start);
         _ = actual.InnerExceptions.Should().NotContain(stop);
 
-        await service.Received(1).StopAsync(Arg.Any<CancellationToken>());
+        await _service.Received(1).StopAsync(Arg.Any<CancellationToken>());
     }
 }